requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
opencv-python==4.8.1.78
numpy==1.24.3
torch==2.1.0
ultralytics==8.0.196
Pillow==10.1.0
vietocr==0.3.1
python-jose==3.3.0 